+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Change Log
 ----------
 
+Development Version (unreleased):
+
+- Add better error messages when operating on a closed file (:issue:`274`, :pull:`275`).
+  By `Kai Mühlbauer <https://github.com/kmuehlbauer>`_
+
 Version 1.6.3 (June 30th, 2025):
 
 - fix invalid string format specifier, match raises/warns with messages in test suite,